backend/ChatAuth/requirements.txt
fastapi 
uvicorn 
requests 
google-auth 
dotenv 
debugpy
PyJWT
pycryptodome 
cryptography
httpx
python-dotenv